Fix bug with buttons toolbar executing commands during playback
Corrects the behaviour when executing context menu commands with the 'Active selection' item group when a track is playing, and selection viewers are set to prefer a currently playing track.

(Previously, nothing would happen when the button is clicked in those circumstances. Now, the current selection is used as was originally intended.)

This problem arose as setting selection viewers to prefer the currently playing track was not originally possible.