Fix toolbars misbehaviour
This fixes some misbehaviour (including a crash bug) in the toolbars due to the difference in behaviour of the RebarBandInfo copy and move constructors (and also the assignment operators).

Some more work is needed to move the run-time data out of RebarBandInfo but these changes will address the problems encountered.