Optimise replacement of all items in playlist view
This uses a more optimised path when all items are being removed from the active playlist.

Flickering is also avoided in this case by some changes in ui_helpers and not calling UpdateWindow in this case. 

(We are, perhaps, being too aggressive in calling UpdateWindow in general. However, this targets the main case causing problems.)