Improve Item properties performance
This significantly improves performance when refreshing the contents of the Item properties panel for large selections.

Refresh times can be as little as 20% of the time taken prior to these changes. This results in a general UI responsiveness improvement, in particular when using library viewers.

Improvements were made partly by removing unneeded processing, and partly by using parallelism.

As a result of these changes, metadata field values are no longer sorted (which was already inconsistent with the rest of foobar2000).

The sort order of track property sections was also corrected.