Make 'Edit this column' command more robust
This fixes a problem where if you were on the playlist view preferences page and had switched away from it to another page, the 'Edit this column' from the playlist view header context menu would not behave correctly.

This is because the code had the expectation that if the preferences dialog had been created, it was visible. However, foobar2000 hides dialogs after switching away from them rather than destroying them. It's noted in a comment elsewhere that this behaviour changed in foobar2000 1.0.