Don't ignore Filter panel selection changes when right-clicking
This changes Filter panel behaviour so that when right-clicking on an item, the selection change is processed normally as it would be when left-clicking.

This means that subsequent Filters are updated, auto-send is triggered and selection viewers are updated.

(Previously, the selection change was effectively ignored, which left the panel in an inconsistent state.)