Rework layout preferences tree item to node object mapping
Previously, a pointer to each node object was stored in the lParam value of each tree view item.

However, this can be prone to error, and there has existed a very rare crash bug relating to this that is of an unknown cause.

This switches to using an unordered map to map tree item IDs to node objects. This is safer, and also simplifies the overall logic.