Partially fix compatibility with latest version of range-v3
Currently, compilation with the latest version of range-v3 is not working.

For Visual Studio, range-v3 now requires the `/experimental:preprocessor` compiler argument. However, specifying this causes an internal compiler error when compiling this project. This will be further investigated separately.

There was also a problem with mixed types when using `ranges::view::iota()`, which was breaking the Clang build (and would probably have also broken the Visual Studio build if it weren't for the other problem). This is fixed here.