Pin ms-gsl and range-v3 ports
There was some breakage with the latest version of `range-v3`. This freezes the ports to working versions to get things working again, and avoid unexpected breakage in future.

Note: The outstanding problem blocking an upgrade of range-v3 is https://developercommunity.visualstudio.com/content/problem/695652/internal-compiler-error-with-experimentalpreproces.html (which is stated to be fixed in Visual Studio 16.4).