Automatically set version number during build using git describe
This adds a pre-build step to automatically set the version number using the output of `git describe` (which generates an identifier based on the latest tag and the distance from it).

The pre-build step executes a PowerShell script which generates a header file containing the version number.

This appears to work well – the script is not executed when no other source files have changed.