Package component following successful builds
This adds a post-build command that creates a .fb2k-component file following a successful build.

The created file is named `foo_ui_columns.fb2k-component`. (This does not include the version number as this complicates the file's removal when cleaning the project.)

(Also, note that `$(TargetDir)` did not work in the 'extensions to delete on clean' setting, and `$(SolutionDir)$(Configuration)` was used as a workaround.)