Fix state bugs in layout tree in layout preferences
This fixes a few bugs in the state management of the layout tree in layout preferences that were leading to misbehaviour and potential crashes.

This happened in particular when reordering panels or removing panels and inserting new ones (after which, the tree items could end up pointing at the wrong nodes).