Add support for multiple field values in in-line editing
This adds support for multiple field values in in-line metadata editing in the playlist view.

This applies to all metadata fields. Multiple values are separated by semicolons.

This also includes a small update to ui_helpers which now detects whether the text has been modified before triggering a save operation.