Fix crash when using inline editing on tracks with varying values
This fixes a crash when saving changes after using inline editing in the playlist view on more than two files with differing field values.

A side effect of the fix is that inline editing is no longer sometimes blocked if a file with no loaded metadata is encountered; this will instead (probably) fail at saving time.