Include version in component archives for releases
This alters the post-build script so that, when a tagged release is being built, the version number is included in the file name of the produced component archive.

For example, if the tag `v1.2.0` is being built, the file name of the archive will be `foo_ui_columns-1.2.0.fb2k-component`.

(For other builds, the file name remains `foo_ui_columns.fb2k-component`.)