Use the Windows Imaging Component to load images
This resurrects the Windows Imaging Component-based image loader, and uses it in the buttons toolbar, artwork panel and playlist view.

This brings support for additional image formats, such as WebP on Windows 10 1809 and newer.

This also adds the Windows Implementation Library as a dependency, and sets _WIN32_WINNT to 0x601 to allow the use of Win32 API functions that require Windows 7.