Improve Item details scroll bar handling
Fixes #331

This rewrites scroll bar-related code in Item details in an effort to improve behaviour when scroll bar information is updated, and fix a problem where it sometimes wasn't possible to scroll to the very bottom if both scroll bars were visible.

This also attempts to preserve the scroll position when changing settings.