Fix Item details incorrect handling of some font changes
#337

This fixes a bug in Item details where font changes that were immediately followed by a colour code weren't correctly handled.

This also adjusts how line heights are calculated, so that if a font change is immediately followed by another one, the first font change is not ignored when calculating the height of the line.