Add support for non-front cover stub images to artwork view
Fixes #314

This adds support for back cover, disc and artist stub images to the artwork view.

These are configured on the main Display page in preferences.

The front cover stub image is used as a fallback if any of these three stub images fail to load (e.g. due to not being configured).