Avoid querying panel configuration data unnecessarily
This reworks the changes that were made in #320 so that the standard splitters no longer refresh a panel's configuration data when `splitter_window::get_panel()` is called.

This is to avoid panel configuration data being unnecessarily queried at strange moments (e.g. when the main window becomes inactive).

Instead, a refresh is done when a panel is being copied, which still fixes the original problem of stale configuration data being copied when a panel is copied during live editing.