Fix inability to create buttons for some File operations commands
Fixes #378

This fixes a problem where it wasn't possible to create a functioning button for the 'File operations/Rename to...' context menu command on current versions of foobar2000, and for other File operations context menu commands on older versions of foobar2000.

Note that any existing non-functioning buttons will need to have the relevant command re-selected in order for the button to function correctly.