Fix misbehaviour when clicking on artwork view panel
#380

This corrects some misbehaviour introduced in #368 when clicking on the artwork view panel to change artwork type.

It also makes a small tweak the the 'Lock artwork type' context menu command so that it always freezes the currently displayed artwork type when the setting is transitioned from disabled to enabled.