Improve behaviour of Reload artwork command
#380

This improves the Artwork view panel 'Reload artwork' context menu command by making it flush any cached image first, to ensure that a new image is always generated and a cached image is not reused.

It also fixes a bug that occurred when using the command, or when changing tracking mode, where dynamic internet artwork artwork was replaced with a stub image.