Make drop-down list toolbars notify host on size limit change
This makes drop-down list toolbars notify the host when size limits change.

This could be due to the configured font changing, or the items in the drop-down list changing.

Note that the toolbars with dynamic items only update their items when the drop-down list is clicked on (or otherwise expanded), as no callbacks are available to update them when the underlying list of items changes. Therefore, the minimum width also only updates on when the control is clicked on.