Update to foobar2000 SDK 2022-01-04 and use vcpkg manifest mode
This updates submodules to bring in changes from foobar2000 SDK 2022-01-04.

Some bits of code were tidied up as part of this and `fmt` was added as a dependency (`std::format()` isn't used as it's not currently available in VS 2022 with the /std:c++20 compiler option).

Finally, vcpkg is now used in manifest mode to simplify handling of dependencies. Currently, only a single manifest in the root directory is used rather than having manifests for each project (I'm not sure how exactly the latter would behave, particularly if the manifests conflicted with each other).