Add dark modes for playlist tabs and tab stack
This adds dark mode support to the playlist tabs and tab stack.

Unfortunately, the Windows tab control does not feature a native dark mode, and its owner-drawn mode is very limited, and so its own painting logic has to be completely overridden.

The left/right buttons when the tab control is in single row mode is still in a light mode.