Add dark mode version of default playback button icons
This adds dark versions of the default icons used for the playback buttons.

These are based on brand new SVG icons, created in a similar style to the existing playback button icons. (New versions of the existing light icons will follow based on these new dark icons.)

The conversion script used to convert the SVGs to ICO files is also included.