Make light and dark mode colour settings independent
This makes the colours configured on the Colours tab of Colours and fonts preferences specific to the active mode (light or dark).

This is so one person can switch between light and dark modes at their will, and have relevant colours configured for each.

On upgrade, any custom colours that are currently active are copied to the dark mode colour settings. This is in case someone already has dark colours configured (which would be otherwise difficult to copy across).

In the colour picker dialogue box shown when changing a custom colour, the custom colour available for selection at the bottom of the dialogue box was also corrected to be the default colour for the currently active mode.

Finally, the Colours tab is now refreshed if the colour mode is set to 'Use system setting' and the system mode changes.