Improve import of legacy spectrum analyser colours
This improves the logic for importing old per-instance spectrum analyser colours into the global unified configuration, so that the import is based on the first actual per-instance colours encountered, rather than the last values that were selected in the configuration dialogue box, as those could differ from the actual colours in some cases (e.g. after an FCL import).