Disable shading of alternate playlist view rows by default
This disables the shading of alternate rows in the playlist view by default in all colour schemes.

Previously, the global style script automatically enabled it for non-themed colour schemes, but it was illogical that wanting to customise colours should also turn this on.

(It can still be enabled by uncommenting the indicated line in the default global style script, however.)