Add dark mode to list view inline editing edit box
This makes the edit box used for inline editing in the various Columns UI list views (playlist view, playlist switcher, filters, item properties, filter preferences) have a dark appearance when dark mode is on.

It also fixes a glitch with the right border of the edit box at high DPIs and prevents glitching during the initial render of the edit box.