Add SVG support to buttons toolbar
This adds SVG support for custom images to the Buttons toolbar using the experimental [SVG services](https://github.com/reupen/svg-services) component. The SVG services component must be installed for this to work.

The SVG services repository has been added as a submodule to access the definition of its public API.