Use SVGs for non-standard icon sizes when possible
This renders built-in icons using SVGs if:

- the display size isn't a size in the corresponding icon resource; and
- the SVG services component is installed

This improves display quality in such cases, as otherwise one of the existing sizes in the icon resource will be scaled.

Pre-rendered icons are still preferred for standard sizes, as they are slightly better quality than what is obtained by rendering the SVG using SVG services.

Standard sizes are square icons that are 16, 20, 24, 28, 32, 36, 40, 48, 56, 64 or 256 pixels in width and height.

This change also includes various bits of related refactoring.