Fix glitches when using tab control spin buttons with dark mode enabled
This resolves some minor glitches that occurred when when using the scroll buttons (spin control) in the Tab stack and Playlist tabs panels when dark mode is enabled.

These were when the hot (hover) status of the tab underneath the spin control was being incorrectly determined when interacting with the spin control, and also when the tabs were being rendered on top of the spin control.