Fix Windows 7 regression in import and export configuration dialogues
This reverts a change in #679 to how the TVS_CHECKBOXES styles was being set in the tree view in the import and export configuration dialogue boxes to fix problems on Windows 7 where the items initial checked state wasn't set correctly.

This is due to how Windows 7 handles that style when it is set in a dialogue box template.