Fix update of maximised title bar when toggling dark mode
This fixes a problem where the title bar of a maximised main window didn't correctly update when switching between light and dark mode.

This was fixed by switching to a different hack to force the title bar to redraw. A workaround is also no longer used on Windows 11, as it doesn't need one.