Fix GDI-object leak in Item properties dialogue box
This fixes a regression caused by #683 on older versions of Windows, where the Item properties options tree view state image list was being leaked every time the dialogue was closed. This is linked to the use of the TVS_CHECKBOXES style.

This change destroys the image list as directed in the documentation. I was only able to reproduce the leak on older versions of Windows (e.g. Windows 7).

(The FCL import and export dialogues already had the logic to destroy the image list, so weren't affected by this problem.)