Make tab control update hover item when using the mouse wheel
This updates the mouse wheel logic in the tab stack and playlist tabs panels so that it makes the tab control update its internal hover item state after scrolling. This is done by sending it a WM_MOUSEMOVE message. This resolves some rendering glitches in both light and dark modes.

(There is some scope to refactor things so that more logic is shared between the two panels, though that is out of scope here.)