Add playlist selector toolbar
This adds a drop-down toolbar for switching between playlists, based on the core `fb2k::dropDownToolbar` interface.

The toolbar name 'Playlist' would be more consistent with other built-in toolbars, but I've made an exception to the usual naming scheme here and added the word 'selector', as 'playlist' on its own could refer to the contents of the playlist.