Make saved main window position DPI aware
This makes the saving and restoring of the main window position between foobar2000 sessions DPI aware. If the system DPI changes between foobar2000 sessions, the position will be scaled accordingly.

(There may be some scope to further improve things by digging into the monitor configuration, but there is a lot of potential complexity there so it's not done here.)