Fix dark mode rendering glitch below tab spin buttons
This fixes a minor rendering glitch below the playlist tabs and tab stack panels when dark mode is enabled. Previously, a tab would render in this small space, but did not re-render when scrolling. Now, tabs are not rendered in that small space and only the background colour shows through,