Work around lag when starting playback of a track
This adjusts the updating of taskbar buttons to use `PostMessage()` rather than a `main_thread_callback`. Why this makes a difference is unclear: the lag is neither when `fb2k::inMainThread()` is called nor when the callback function is called. The reason may be related to `ITaskbarList3::ThumbBarUpdateButtons` calling `SendMessageTimeout` (without `SMTO_BLOCK`) which was previously noted as being the cause of unexpected problems.

(Doing the taskbar button updates off the main thread is an option, but I don't believe that should be needed with this change.)