Save DirectWrite font family, weight, stretch, style and DIP size
This makes changes to how fonts selected in the Colours and fonts preferences page are saved so that DirectWrite-compatible properties are retained and the selected DirectWrite font can be used by panels.

A preliminary version of a `cui::fonts::manager_v3` interface was added in order to allow panels to access the extended font configuration.

Note that none of this yet fully supports the [typographic font family model](https://learn.microsoft.com/en-us/windows/win32/directwrite/font-selection#typographic-font-family-model) and variable font axes.