Store extended font configuration in backwards-compatible way
Unfortunately, due to a bug in the font configuration data reader, downgrading to the release version of Columns UI was causing incorrect custom font sizes to be read.

This reworks how the font configuration data is written so that things remain backwards compatible with the release version of Columns UI.

Note that the extended DirectWrite-specific font configuration data written by recent development versions of Columns UI won't be read by this version due to the change in how the configuration data is written (and therefore font configuration will be reimported from the old GDI-based configuration structures).