Allow configuration of some DirectWrite rendering options
This allows the configuration of the DirectWrite rendering mode, and whether to force the use of greyscale anti-aliasing.

These options apply to built-in parts of the main UI that use DirectWrite for text rendering, and will also be exposed in the updated font API.

Hopefully these options should be enough to keep anyone that doesn't like the default appearance of DirectWrite text happy.