Add new $set_format() title formatting function to Item details
This adds a new title formatting function, `$set_format()`, to Item details.

This is a replacement for the older `$set_font()` function.

The new function takes a single argument, which is a CSS-like string. For example:

```
$set_format(
  font-family: Segoe UI Variable;
  font-size: 20;
  font-weight: 300;
  font-stretch: 5;
  font-style: italic;
  text-decoration: underline;
)
```

Properties can be omitted (in which case they will retain their current value).

A special value of `initial` can be used to reset any particular property to its default value.

The `$reset_format()` function can be used to reset all properties to their defaults in one go.