Add workflow to auto-merge dependency update pull requests
This adds, as a trial, a workflow to automatically merge Dependabot and pre-commit pull requests.

This is based on the example on https://docs.github.com/en/code-security/dependabot/working-with-dependabot/automating-dependabot-with-github-actions.

Pull requests with more than 1 commit are ignored, and for pre-commit PRs there is an additional check that only `.pre-commit-config.yaml` has been modified.