Add support for percentage font-stretch values in Item details
This adds support for percentage `font-stretch` values when using `$set_format()` in Item details.

Note that a `%` character needs to be escaped in title formatting, therefore percentages for `font-stretch` need to be written as e.g. `100%%` or `100pc`.

The percentage values are most useful on recent versions of Windows where variable font axes are supported.