Fix rare drag-and-drop crash
This should fix a rare drag-and-drop crash when dropping items on certain parts of the main window (and probably also the playlist tabs on reviewing the code).

This was related to some incorrect usage of `m_DataObject` in the `IDropTarget::Drop` implementation in a `uih::ole::set_drop_description()` call (and somehow `m_DataObject` being empty at the same time).

I could not recall a reason for removing the drop description in the `IDropTarget::Drop` implementation. It was not being done consistently, so I've stopped doing it everywhere. (Additionally, there was an incorrect call to `IDropTargetHelper::DragLeave()`...