Fix incorrect font family use in DirectWrite panels on Windows 10
This corrects an assumption when creating DirectWrite text formats that if axis values are populated then the typographic font family name is also populated.

This assumption was incorrect on most versions of Windows 10 and older versions of Windows 11.

It now falls back to the weight-stretch-style font family name if the typographic font family name is missing and axis values are populated.

Additionally, a few other minor tweaks were made in `ui_helpers` (namely to handle simulated fonts correctly on the same Windows versions as above).