Fix dark mode handling on Fonts preferences tab on foobar2000 1.x
This fixes a minor bug where the owner drawn combo boxes on the Fonts tab of the Colours and fonts preferences page incorrectly had a dark appearance in some cases on on foobar2000 1.x (which does not support dark mode for Preferences). This happened when Columns UI was the active UI and had dark mode enabled.