Fix incorrect playlist view header positioning after certain actions
This fixes a problem where, in a few scenarios, the playlist view header could be incorrectly positioned.

This happened, for example, when the playlist view was scrolled right and auto-sizing columns was turned on or items were removed from the playlist causing the vertical scroll bar to disappear.

Additionally, a problem where the playlist vertical scroll position was sometimes reset after showing or hiding columns via the header context menu was fixed.