Track changed artwork in Artwork view correctly on aborted artwork loads
When e.g. holding down the Up or Down keys in the playlist view and causing many requests to load artwork, the Artwork view was sometimes losing track of whether the artwork had changed between tracks. (This happened after a read was aborted due to the tracked track change.)

This tweaks the logic to avoid that.