Pin C++/WinRT headers to last version compatible with Windows 7
The May Windows SDK update (version 10.0.26100.4188) includes an update to the `cppwinrt` headers that breaks compatibility with Windows 7 (and also stops Columns UI from building as it tries to link against those functions not available on Windows 7).

This switches to installing `cppwinrt` via vcpkg and pinning it to version 2.0.230706.1, the last one compatible with Windows 7. This version was not available via vcpkg, so an overlay port had to be used.