Fix crash after unexpected exception reading artwork in Artwork view
Some input components throw unusual exceptions which are then exposed through `album_art_manager_v2` methods.

In this scenario, previously reading the stub image was skipped, but the component still tried to display it and ended up crashing if it hadn’t been read in the past.

This updates the logic to avoid that. Also, the panel now displays nothing rather than a stub image when such an error occurs.