Add Windows Advanced Color support to Artwork view
This adds support for [Advanced Color](https://learn.microsoft.com/en-gb/windows/win32/wcs/advanced-color-icc-profiles), available in Windows 11 and later versions of Windows 10, to the Artwork view.

This improves support for embedded colour profiles, HDR images and high-bit depth images on supported versions of Windows.

Note that the documentation is a bit confusing in part, and the functionality is a bit more limited for Win32 apps (compared to UWP ones), so some more checks will need to need to be done to check that the colour management is working correctly.

Support for legacy colour management will probably be added separately as well (it will be needed for artwork in the playlist view anyway).