Update spectrum analyser logarithmic frequency scale
This updates the spectrum analyser visualisation to use the same frequency scale as the Default UI Spectrum visualisation when the x-axis is in logarithmic mode.

This means a frequency range of 50 Hz to 25,000 Hz, and the scale is standardised for different sample rates.

Some configuration for a few parameters will probably be added later.

A few bits of code were also cleaned up.