Use Direct2D to prerender playlist view artwork
This migrates playlist view artwork scaling and reflection rendering from GDI+ to Direct2D.

A memory bitmap is still used and the prerendered image is still converted to a GDI bitmap. However, the GDI bitmap now uses premultiplied alpha and is alpha blended on to the GDI device context.

Additionally, a problem where artwork images in the playlist view varied in width by up to two pixels was also fixed.

Also, in some cases, the transparency gradient in reflections is slightly different now. As far as I can determine, the new behaviour is correct.