Add basic colour management when rendering playlist view artwork images
This adds basic (legacy) colour management support for artwork images shown in the playlist view.

Legacy colour management is normally based on using display colour profiles.

If ‘auto colour management’ is enabled in Windows, however, Windows will normally pretend that the display has no colour profile to applications, to make them target sRGB.

The behaviour is further modified when ‘Use legacy display ICC colour management’ is enabled in the Compatibility settings for an application (while ‘auto colour management’ is on). In this case, Windows will return a synthetic colour profile (rather than no profile) so the full gamut of the display can be used.

For more info, see https://learn.microsoft.com/en-gb/windows/win32/wcs/advanced-color-icc-profiles.

Note that, for Win32 apps, there seems to be no notifications when display colour profiles are changed, and so foobar2000 needs to be restarted when display colour profiles are modified in Windows.