Add support for legacy colour management to artwork view
This adds support for basic (legacy) colour management to the artwork view, and makes Advanced Colour opt-in via Preferences.

Legacy colour management is now the default, as it works on all supported versions of Windows, and Advanced Colour requires the flip-model for the swap chain which causes artefacts when resizing the artwork view. Additionally, Advanced Colour causes colour differences when HDR isn't enabled in Windows.