Update layout tree context menu wording and order
This makes a few tweaks to the wording of context menu items of the layout tree in Preferences.

It also moves the previous ‘Change base’ item down, as having it first was asking for trouble (since it replaces the root item and all children).

Finally, the current splitter type is now indicated in the ‘Splitter type’ submenu.