Show stub image correctly after settings are changed
Previously, the stub was shown correct after track changes or changing the tracking mode. However, it wasn’t shown in some other scenarios such as after enabling ‘Lock artwork type’ or after a Direct2D device reset.

This amends the logic for showing it so it’s always shown when it should be.