Improve handling of panel removal during panel creation in Tab stack
This fixes a specific problem during live editing where it was possible to insert a single-instance panel twice into a Tab stack, but this would then crash.

This was due to the first instance of the panel being removed while the panels were being iterated over.

It also improves handling of removing panels during live editing in the Tab stack.