Add move to previous and next position live editing commands
This adds a ‘Move to’ submenu to the live editing context menu that contains commands for moving a panel to the previous or next position in its parent splitter.

Some bugs in the Tab stack handling of live panel replacement were also fixed.

(Still some room for further improvement for how reordering is handled, though will have to come back to that.)