Fix space digit grouping symbol in status bar selected track count
This fixes a bug where, if a non-breaking space was configured to be the digit grouping symbol in Windows, the selected track count in the status bar was truncated at the first instance of the digit grouping symbol.

This was because the character injected from the locale by `fmt::format()` wasn't UTF-8-encoded.

This change switches to using UTF-16 for strings in the status bar to fix the problem.