Add internal clipboard when cutting, copying and pasting panels
This adds a secondary internal clipboard when cutting, copying and pasting panels during live editing or in the layout tree in preferences.

When pasting, the internal clipboard is used as a fallback if a panel isn’t found in the Windows clipboard. This is to protect against accidentally overwriting the Windows clipboard with something else (which is particularly problematic now that you can cut panels).

The internal clipboard is only held in memory, so it doesn’t persist once the process is closed.