Allow copying and removing root panel during live editing
This makes it possible to cut, copy and remove the root panel during live editing.

When removing, confirmation is requested if the root panel is a splitter.

After removing the root panel, it’s possible to paste a panel or add a new root panel.

The ‘Replace base’ submenu has been removed as it’s no longer needed with these changes.