Allow removing the root panel in Preferences layout tree
This makes it possible to remove the root panel from the Preferences layout tree.

When removing, confirmation is requested if the root panel is a splitter.

After removing the root panel, it’s possible to paste a panel or add a new root panel.

The ‘Replace base’ submenu has been removed as it’s no longer needed with these changes.