Adjust artwork view rendering intent when Advanced Colour is enabled
According to the [Microsoft documentation](https://learn.microsoft.com/en-gb/windows/win32/wcs/advanced-color-icc-profiles), “When you're in an advanced color workflow, we generally don't recommend using the perceptual rendering intent, neither for source or destination, since it was designed for SDR sources and destinations that have smaller color gamuts than the ones used for HDR and some WCG displays; so using them can result in unexpected behavior.”

Helpfully, it doesn’t explain what to actually do.

Relative colorimetric seems to get correct results on an HDR display, or when auto colour management is enabled in Windows, so this switches to that (when Advanced Colour is enabled in Artwork view).