Remove shown image in Artwork view for playing track if artwork is removed
This fixes a bug where the Artwork view didn’t update if it was showing the front cover for the playing track, and then the foobar2000 now-playing album art manager reported that the artwork was removed.

It also fixes some unnecessary work that was happening when non-front-cover art was shown for the playing track.