Fix documentation link-related Setup preferences tab problems
The solution for changing the dark mode link colour wasn’t working on foobar2000 2.24. This fixes that by updating `fbh` to make the subclassing happen earlier. The link was also moved down further from the heading above it.

This also fixes a bug where other controls on the tab triggered a crash, due to a missing `break` under `WM_COMMAND`.