Prevent toolbar callbacks from being released during process exit
In some types of abnormal process exits, apparently destructors for objects with static storage duration get called. If the event that caused the abormal exit was on a non-main thread, this then means those destructors get called on that thread.

In the case of the Output device and Output format toolbars, a single static callback object was created for all instances of the toolbars. In the normal scenario it gets released when the last instance is destroyed. But in the bad scenario, the abnormal process exit is triggered from another thread, and the destructors for those callback objects cause further problems as they are only valid to be called from the main thread.

This simply changes those toolbars to wrap the callback object and not bother releasing it in the destructor to avoid this. (In practice, how much this will help I don’t know, as there must be all kinds of destructors being called in this scenario.)