Generate UTF-8 messages for WIL exceptions before printing them
This addresses a potential problem that has been lurking around for a while. `wil::ResultException::what()` returns an ANSI string, which was then printed to the console in some places.

This changes places where `wil::ResultException` may have been logged in that way to use a new helper function in `mmh`. That helper function returns a UTF-8 version of the message for `wil::ResultException` exceptions, and returns a copy of the original message for any other `std::exception`.