Update alternative emoji font selection logic emoji list
This updates ui_helpers to pick up an update to the emoji list used for the ‘Use alternative emoji font selection logic’ setting.

The emoji list has been updated from the previous Unicode 17 draft to the final version of Unicode 17.0.0. The only change is the removal of an apple core emoji.