Fix rendering of some non-RGB images
This resolves a problem were some CMYK and high-bit-depth greyscale images did not render correctly.

It seems like `WICConvertBitmapSource`/`IWICFormatConverter` converts these to sRGB automatically (at least, I was unable to see any noticeable difference from doing a manual conversion).

Doing an explicit conversion to sRGB is also a bit of a pain as it can fail in cases, so this change works on the assumption that `IWICFormatConverter` is doing it.