Fix loading of display ICC profiles with embedded WCS profile
It turns out that there are ICC profiles that contain embedded Windows Color System (WCS) profiles. An example of these is profiles created by the built-in calibration tool in Windows. `WcsOpenColorProfile` returns the WCS profile in these cases unless a flag is set, and so this sets that flag to get the ICC profile, and fix the onward conversion of these profiles into Direct2D colour contexts.

It seems Windows these WCS profiles can also exist in a standalone format with the .cdmp file extension: https://learn.microsoft.com/en-gb/windows/win32/wcs/wcs-color-device-model-profile-schema-and-algorithms. These should be ignored as the call to `WcsGetDefaultColorProfile()` requests an ICC profile (though this should be tested).