Ignore Microsoft CDMP display profiles
This makes the display colour profile reader ignore Microsoft CDMP profiles.

As far as I can see, CDMP profiles have little effect on software on Windows 11. The main thing that was noticeably affected by one was Photoshop.

The classic Windows Picture Viewer is also affected by them. However, the Photos app does not appear to be.

It is fairly straightforward to support them via `WcsCreateIccProfile`. But the aim is to be consistent with other software (and in particular the Photos app), so this just ignores them.