Ignore CMYK ICC profiles embedded in RGB images
This makes the Artwork view and playlist view ignore CMYK colour profiles embedded in RGB images. This can apparently happen in PNG images that have been badly converted from JPEGs.

Unfortunately, the ICC colour profile signature doesn’t seem to be exposed by WIC, so this reads into the profile bytes to get it. WCS could be used to read the profile instead, though it’s not clear there be much benefit for getting this single value. A full list of signature values can be found in the [ICC specification](https://www.color.org/specification/ICC.1-2022-05.pdf) (page 21 for that version of the spec).