Handle DXGI occlusion status correctly in Artwork view
This adds missing handling for the `DXGI_STATUS_OCCLUDED` return value of `IDXGISwapChain::Present()`.

This status occurred when the monitor was in sleep mode or off, the desktop was locked, or a UAC prompt was active. Sometimes, this would result in the Artwork view showing an old image when the window became visible again.

On Windows 8 and newer, `IDXGIFactory2::RegisterOcclusionStatusWindow()` is used to receive a notification when the window is no longer occluded.

Unfortunately, that method is not supported on Windows 7, so it falls back to using a timer there.