Add Focus Filter search toolbar menu item
This adds a hidden menu item for focusing the Filter search toolbar.

This is designed for use as a keyboard shortcut. Note that using the F3 key when a Filter panel is focused was already supported.

When opening the View menu, the new command will only appear if the Shift key was held down.