Ignore diacritics when typing in list views to jump to an item
This makes jumping to items by typing in built-in list views ignore diacritics and symbols. Additionally, full and half-width characters are now considered the same, and there will be other behaviour changes due to the change in comparison function used and the flags specified.

These changes were made in the ui_helpers and mmh libraries.